Add Attention Folder and Functions, and Change Self Attention and Cross Attention in Decoder Class
